requirements.txt
python-dotenv==1.0.0
selenium==4.11.2
tenacity~=8.2.3